+++ 2024/20241201/camomile.kr/index.ts
@@ -1,0 +1,5 @@
+import { Puzzle_P20241201 } from './part1'
+
+const puzzle = new Puzzle_P20241201();
+
+puzzle.solvePuzzle();
