+++ 2024/BaseSolvePuzzle.ts
@@ -1,0 +1,3 @@
+export interface BaseSolvePuzzle {
+    solvePuzzle();
+}
